Bugfix: 1385: Authentication fails if both date and both x-ms-date isiF specified
For queues and blobs, when forming the shared key token, the Date header
should be treated as the empty string when the x-ms-date header is
present.

For tables, the value of the x-ms-date header shoudl be used as the
value of the Date header if teh x-ms-date header is present.

This behavior is documented at
https://learn.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key