Activate only on command or Azurite file presence.
Optionally add `onStartupFinished`, however if you're going to do that, no point in having any of the other activation events and they should all be removed.

Potentially considered a breaking change because a "new" workspace won't activate the extension and people may be used to seeing the blob/table/etc. links in the bottom status tray and they won't show up until you run an azurite command. Once run however and the azurite files are present in the workspace, it will auto-start from then on for the project.